Extend abstraction for mutliple table types
- abstract base class for renderer factory
- generalize TableColumn and RendererTable
- move specific functions into derived factory renderer class
