+++ rg_ad_races_edit.inc.php
@@ -61,7 +61,7 @@ while ($zaznam = mysqli_fetch_array($vysledek, MYSQLI_ASSOC)) {
 $renderer_option['curr_date'] = GetCurrentDate();
 
 // define table
-$tbl_renderer = new RacesRenderedTable();
+$tbl_renderer = RacesRendererFactory::createTable();
 $tbl_renderer->addColumns('datum','nazev','misto','oddil');
 if ($ext_id_active_oris)
 	$tbl_renderer->addColumns('ext_id');
